+++ hydromt_wflow/workflows/__init__.py
@@ -3,6 +3,7 @@
 from .basemaps import *
 from .connect import *
 from .demand import *
+from .forcing import *
 from .gauges import *
 from .glaciers import *
 from .landuse import *
@@ -10,4 +11,5 @@ from .river import *
 from .rootzoneclim import *
 from .soilgrids import *
 from .soilparams import *
+from .states import *
 from .waterbodies import *