+++ docs/changelog.rst
@@ -13,7 +13,11 @@ Added
 -----
 - New function **setup_1dmodel_connection** to connect wflow to 1D river model (eg Delft3D FM 1D, HEC-RAS, etc.) `PR #210 <https://github.com/Deltares/hydromt_wflow/pull/210>`_
 - New setup method for the **KsatHorFrac** parameter **setup_ksathorfarc** to up-downscale existing ksathorfrac maps. `PR #255 <https://github.com/Deltares/hydromt_wflow/pull/255>`_
+- new function **setup_pet_forcing** to reproject existing pet data rather than computing from other meteo data. PR #257
 - Workflow to compute brooks corey c for the wflow layers based on soilgrids data, soilgrids_brooks_corey. PR #242
+- better support for WflowModel states with new methods: **read_states**, **write_states** and **clip_states**. PR #252
+- new function **setup_cold_states** to prepare cold states for WflowModel. PR #252
+- new utils method **get_grid_from_config** to get the right wflow staticmaps variable based on the TOML configuration (e.g. detects name in netcdf, value, scale and offset). Only applied now to prepare cold states (e.g. not yet in read_grid). PR #252
 
 Changed
 -------
@@ -23,6 +27,8 @@ Changed
 Fixed
 -----
 - Wrong dtype for wflow_subcatch map. PR #247
+- **setup_gauges**: Allow snapping to river/mask for snap_uparea method. PR #248
+- Removed building a wflow model without a config file in the build notebook.
 
 v0.5.0 (February 2024)
 ======================